chartsy version 1.2.0

+++ Chartsy/Main/src/org/chartsy/main/utils/FileUtils.java
@@ -3,6 +3,7 @@ package org.chartsy.main.utils;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.FilenameFilter;
 import java.io.IOException;
 import java.nio.channels.FileChannel;
 import org.openide.filesystems.FileObject;
@@ -56,13 +57,43 @@ public final class FileUtils {
         return result;
     }
 
+    public static String cacheFolder()
+    {
+        String result = LocalFolder() + File.separator + "cache";
+        createFolder(result);
+        return result;
+    }
+
+    public static String cacheFile(String file)
+    {
+        String result = getFileName(cacheFolder(), file);
+        createFile(result);
+        return result;
+    }
+
+    public static String getHistoryFolder()
+    {
+        String result = LocalFolder() + File.separator + "history";
+        createFolder(result);
+        return result;
+    }
+
+    public static boolean fileExists(String path)
+    {
+        File file = new File(path);
+        return file.exists();
+    }
+
     public static void removeFile(String path) {
         File file = new File(path);
-        try {
-            FileObject fo = FileUtil.toFileObject(file);
-            fo.delete();
-        } catch (IOException ex) {
-            ex.printStackTrace();
+        if (file.exists())
+        {
+            try {
+                FileObject fo = FileUtil.toFileObject(file);
+                fo.delete();
+            } catch (IOException ex) {
+                ex.printStackTrace();
+            }
         }
     }
 
@@ -78,6 +109,12 @@ public final class FileUtils {
         catch (IOException ex) { ex.printStackTrace(); }
     }
 
+    public static void copyFile(String source, String destination) throws IOException {
+        File sourceFile = new File(source);
+        File destinationFile = new File(destination);
+        copyFile(sourceFile, destinationFile);
+    }
+
     public static void copyFile(File source, File destination) throws IOException {
         FileChannel in = null;
         FileChannel out = null;
@@ -89,6 +126,23 @@ public final class FileUtils {
             if (in != null) in.close();
             if (out != null) out.close();
         }
+    }
+
+    public static String getFileName(String folder, final String path)
+    {
+        File dir = new File(folder);
+        String[] list = dir.list(new FilenameFilter()
+        {
+            public boolean accept(File dir, String name)
+            {
+                return name.contains(path);
+            }
+        });
+        
+        if (list.length == 0)
+            return folder + File.separator + path;
+        else
+            return folder + File.separator + path + "(" + list.length + ")";
     }
 
 }
